refactor: change api/v1/ folder structure
