+++ front/src/pages/Admin/Home/Home.jsx
@@ -14,10 +14,13 @@ import { useTranslation } from 'react-i18next';
 import { useHistory, useLocation } from 'react-router-dom';
 import { useQuery } from 'react-query';
 import DebouncedSearch from '@sb-ui/components/atoms/DebouncedSearch';
-import { appointTeacher } from '@sb-ui/utils/api/v1/user';
+import {
+  appointTeacher,
+  getUsers,
+  removeTeacher,
+} from '@sb-ui/utils/api/v1/admin';
 import { getQueryPage } from '@sb-ui/utils/utils';
 import { ADMIN_USERS_BASE_KEY } from '@sb-ui/utils/queries';
-import { getUsers, removeTeacher } from '@sb-ui/utils/api/v1/user/user';
 import { MainDiv, TableHeader, TitleHeader } from './Home.styled';
 
 const messageKey = 'teacherStateLoading';
