+++ front/src/pages/User/EnrollModal/EnrollModal.desktop.jsx
@@ -3,10 +3,7 @@ import { useHistory, useLocation, useParams } from 'react-router-dom';
 import { useQuery } from 'react-query';
 import { useTranslation } from 'react-i18next';
 import { Col, Divider, Modal, Rate, Typography } from 'antd';
-import {
-  getEnrolledLesson,
-  postEnroll,
-} from '@sb-ui/utils/api/v1/lesson/lesson';
+import { getEnrolledLesson, postEnroll } from '@sb-ui/utils/api/v1/student';
 import { LESSON_PAGE, USER_HOME } from '@sb-ui/utils/paths';
 import { USER_LESSON_MODAL_BASE_KEY } from '@sb-ui/utils/queries';
 import lessonImg from '@sb-ui/resources/img/lesson.svg';
