+++ front/src/pages/User/Home/OngoingLessons/LessonsList/useLessonsRequest.js
@@ -1,5 +1,5 @@
 import { useEffect, useState } from 'react';
-import { getEnrolledLessons } from '../../../../../utils/api/v1/lesson/lesson';
+import { getEnrolledLessons } from '@sb-ui/utils/api/v1/student';
 
 export const useLessonsRequest = () => {
   const [loading, setLoading] = useState(false);
