+++ front/src/pages/User/Home/OngoingLessons/OngoingLessons.jsx
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 import { useQuery } from 'react-query';
 import MobileContext from '@sb-ui/contexts/MobileContext';
 import { USER_ENROLLED_SHORT_LESSONS_BASE_KEY } from '@sb-ui/utils/queries';
-import { getEnrolledLessons } from '@sb-ui/utils/api/v1/lesson';
+import { getEnrolledLessons } from '@sb-ui/utils/api/v1/student';
 import OngoingLessonsMobile from './OngoingLessons.mobile';
 import OngoingLessonsDesktop from './OngoingLessons.desktop';
 import { PAGE_SHORT_SIZE } from '../PublicLessons/constants';
