+++ front/src/utils/api/v1/admin.js
@@ -2,36 +2,6 @@ import api from '@sb-ui/utils/api';
 
 const PATH = '/api/v1/user';
 
-export const postSignUp = async (formData) => {
-  try {
-    const { status, data } = await api.post(`${PATH}/signup`, {
-      ...formData,
-    });
-    return { status, data };
-  } catch (e) {
-    const { status, data } = e.response;
-    return {
-      status,
-      data,
-    };
-  }
-};
-
-export const postSignIn = async (formData) => {
-  try {
-    const { status, data } = await api.post(`${PATH}/signin`, {
-      ...formData,
-    });
-    return { status, data };
-  } catch (e) {
-    const { status, data } = e.response;
-    return {
-      status,
-      data,
-    };
-  }
-};
-
 export const getUsers = async ({ queryKey }) => {
   const [, paramsData] = queryKey;
 
@@ -65,23 +35,6 @@ export const removeTeacher = async (id) => {
   try {
     const { status, data } = await api.post(`${PATH}/remove_teacher`, { id });
     return { status, data };
-  } catch (e) {
-    const { status, data } = e.response;
-    return {
-      status,
-      data,
-    };
-  }
-};
-
-export const getUser = async () => {
-  try {
-    const { status, data } = await api.get(`${PATH}/self`);
-
-    return {
-      status,
-      data,
-    };
   } catch (e) {
     const { status, data } = e.response;
     return {
