--- front/src/utils/api/v1/lesson/index.js
@@ -1,1 +1,0 @@
-export * from './lesson';
