--- front/src/utils/api/v1/lesson/lesson.js
@@ -1,130 +1,0 @@
-import api from '@sb-ui/utils/api';
-
-const PATH = '/api/v1/lesson';
-
-export const getLessons = async (paramsData) => {
-  try {
-    const { status, data } = await api.get(`${PATH}/`, {
-      params: paramsData,
-    });
-    return { status, data };
-  } catch (e) {
-    const { status, data } = e.response;
-    return {
-      status,
-      data,
-    };
-  }
-};
-
-export const getLessonById = async ({ queryKey }) => {
-  const [, paramsData] = queryKey;
-  const { id } = paramsData;
-  const { data } = await api.get(`${PATH}/${id}`);
-
-  return data;
-};
-
-export const postLessonById = async (paramsData) => {
-  const { data } = await api.post(`${PATH}/${paramsData.id}/learn`, paramsData);
-
-  return data;
-};
-
-export const postEnroll = async (id) => {
-  try {
-    const { status, data } = await api.post(`${PATH}/enroll/${id}`, {});
-    return { status, data };
-  } catch (e) {
-    const { status, data } = e.response;
-    return {
-      status,
-      data,
-    };
-  }
-};
-
-export const getEnrolledLessons = async ({ queryKey }) => {
-  const [, paramsData] = queryKey;
-
-  const { data } = await api.get(`${PATH}/enrolled/`, {
-    params: paramsData,
-  });
-
-  return data;
-};
-
-export const getEnrolledLessonsFinished = async ({ queryKey }) => {
-  const [, paramsData] = queryKey;
-
-  const { data } = await api.get(`${PATH}/enrolled-finished/`, {
-    params: paramsData,
-  });
-
-  return data;
-};
-
-export const getEnrolledLesson = async ({ queryKey }) => {
-  const [, paramsData] = queryKey;
-
-  const { data } = await api.get(`${PATH}/enroll/${paramsData.id}`);
-
-  return data;
-};
-
-export const getLesson = async ({ queryKey }) => {
-  const [, { id }] = queryKey;
-
-  const { data } = await api.get(`${PATH}/maintain/${id}`);
-  return data;
-};
-
-export const putLesson = async (params) => {
-  const { data } = await api.put(
-    `${PATH}/maintain/${params.lesson.id}`,
-    params,
-  );
-  return data;
-};
-
-export const createLesson = async (values) => {
-  const { data } = await api.post(`${PATH}/maintain/`, values);
-  return data;
-};
-
-export const getTeacherLessons = async ({ queryKey }) => {
-  const [, paramsData] = queryKey;
-
-  const { data } = await api.get(`${PATH}/maintain/`, {
-    params: paramsData,
-  });
-
-  return data;
-};
-
-export const getTeacherStudents = async ({ queryKey }) => {
-  const [, paramsData] = queryKey;
-
-  const { data } = await api.get(`${PATH}/maintain/students`, {
-    params: paramsData,
-  });
-
-  return data;
-};
-
-export const getPublicLessons = async ({ queryKey }) => {
-  const [, paramsData] = queryKey;
-
-  const { data } = await api.get(`${PATH}/`, {
-    params: paramsData,
-  });
-  return data;
-};
-
-export const getUserLesson = async ({ queryKey }) => {
-  const [, id] = queryKey;
-
-  const { data } = await api.get(`${PATH}/${id}`);
-
-  return data;
-};
