--- front/src/utils/api/v1/user/index.js
@@ -1,1 +1,0 @@
-export * from './user';
